refactor(slice-45): move sweep orchestration into server.core.pipeline
Group orchestrator, executors, experiment_control, and startup_reconciliation
under the Behavior-theme pipeline package; alias old import paths so existing
orchestrator patch targets keep working while callers move to canonical imports.
