+++ docs/plan/DECISIONS.md
@@ -151,3 +151,4 @@
 | 145 | 2026-07-27 | review | nw-solution-architect-reviewer Slice 45 roadmap iter2 → **APPROVED** (0 blocking) | Post–Gap 8 + resume; #137 remediations still hold; Before-Checks live count **1/10** (orchestrator corrected reviewer’s 8/11 claim) | NEEDS_REVISION / re-open Must move tables |
 | 146 | 2026-07-27 | 45 | Phase 1 skateboard = `core/guards/` then `pipeline/` orchestrator SLAP (not whole `core/` in one PR) | Blast-radius core ≈112 match-lines (&lt;200); CLI touches `search_index_*`; highest cognitive load still orchestrator — sequence guards first for safer import rewrite | Move entire `server/core/` in one PR / FE-primitives-first |
 | 147 | 2026-07-27 | 45 | Landed `server/core/guards/` + shim re-exports at old `server.core.*` paths; test patches retargeted to `server.core.guards.*` | Preserve caller imports for one release; mypy-safe explicit re-exports; coverage cov paths → guards | sys.modules alias shims / rewrite all callers in same PR |
+| 148 | 2026-07-27 | 45 | Land `server/core/pipeline/` (orchestrator/executors/experiment_control/startup_reconciliation); `sys.modules` shims for patch compatibility | Mega-suite patches `server.core.orchestrator.*`; function-level SLAP deferred to next PR on same branch | Split `_run_single` / bayesian in same commit as package move |
