+++ server/core/pipeline/__init__.py
@@ -1,0 +1,5 @@
+"""Pipeline / sweep orchestration — Behavior theme (Slice 45).
+
+Canonical imports live under ``server.core.pipeline``. Temporary shims remain at
+``server.core.<module>`` for one release (see CHANGELOG Deprecated).
+"""
